backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
psutil==5.9.8
apscheduler==3.10.4
openai==1.50.0
python-dotenv==1.0.1
